requirements.txt
requests==2.32.3
tabulate==0.9.0
python-dotenv==1.1.0